% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{cfl_games}
\alias{cfl_games}
\title{Query for CFL game information}
\usage{
cfl_games(season = NA, game_id = NA)
}
\arguments{
\item{season}{Required passed parameter for the football season.}

\item{game_id}{Optional passed parameter for the football game id.}
}
\value{
Returns a tibble with game information
}
\description{
Query for CFL game information
}
\examples{
\dontrun{
cfl_games(season = 2016, game_id = 2280)
}
}
