% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{cfl_players}
\alias{cfl_players}
\title{Query for CFL player information}
\usage{
cfl_players(player_id = NA, type = "meta")
}
\arguments{
\item{player_id}{Required parameter for the CFL player ID.}

\item{type}{Future parameter for the type of player lookup (ie. general, game data, season summary)}
}
\value{
Returns a tibble with game information
}
\description{
Query for CFL player information
}
\examples{
\dontrun{
cfl_players(138985)
}
}
