% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_pbp.R
\name{cfl_plays}
\alias{cfl_plays}
\title{Query for CFL Play-by_Play information}
\usage{
cfl_plays(season = NA, game_id = NA)
}
\arguments{
\item{season}{Required parameter for the football season.}

\item{game_id}{Required parameter for the football game id.}
}
\value{
Returns a tibble with play information for a game
}
\description{
Query for CFL Play-by_Play information
}
\examples{
\dontrun{
cfl_plays(season = 2016, game_id = 2280)
}
}
