% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCFL.R
\docType{package}
\name{tidyCFL-package}
\alias{tidyCFL}
\alias{tidyCFL-package}
\title{tidyCFL: tidydata R interface for the CFL API}
\description{
A tidydata package for the official Canadian Football League API
}
\details{
The premier R package for retreiving data from the Official Canadian Football League (CFL) API.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.github.com/andrewjdyck/tidyCFL}
  \item Report bugs at \url{http://www.github.com/andrewjdyck/tidyCFL/issues/new}
}

}
\author{
\strong{Maintainer}: Andrew Dyck \email{andrew@andrewdyck.com}

Authors:
\itemize{
  \item Dan Fleischeker
}

}
\keyword{CFL}
\keyword{Canadian}
\keyword{Footbal}
\keyword{League}
\keyword{internal}
