% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCFLconfig.R
\name{tidyCFL.api_key}
\alias{tidyCFL.api_key}
\title{Query or set CFL API key}
\usage{
tidyCFL.api_key(api_key)
}
\arguments{
\item{api_key}{Required passed parameter to set CFL \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}.
}
\description{
Query or set CFL API key
}
\examples{
\dontrun{
tidyCFL.api_key('foobar')
}
}
